requirements.txt

arango
python-dotenv
python-arango
requests
tqdm
sentence-transformers
hf_xet
flask
numpy
